Stream Api and Lambda

+++ src/main/java/com/tobeto/rentACar/services/concretes/CustomerManager.java
@@ -3,10 +3,15 @@ package com.tobeto.rentACar.services.concretes;
 import com.tobeto.rentACar.entities.Customer;
 import com.tobeto.rentACar.repositories.CustomerRepository;
 import com.tobeto.rentACar.services.abstracts.CustomerService;
+import com.tobeto.rentACar.services.dtos.brand.responses.GetListBrandResponse;
 import com.tobeto.rentACar.services.dtos.customer.requests.AddCustomerRequest;
 import com.tobeto.rentACar.services.dtos.customer.requests.DeleteCustomerRequest;
 import com.tobeto.rentACar.services.dtos.customer.requests.UpdateCustomerRequest;
+import com.tobeto.rentACar.services.dtos.customer.responses.GetListCustomerResponse;
+import com.tobeto.rentACar.services.dtos.customer.responses.GetListValidLicenceCostumerResponse;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class CustomerManager implements CustomerService {
@@ -20,8 +25,8 @@ public class CustomerManager implements CustomerService {
     @Override
     public void add(AddCustomerRequest request) {
         Customer customer = new Customer();
-        customer.setFirstName(request.getFirst_name());
-        customer.setLastName(request.getLast_name());
+        customer.setFirstName(request.getFirstName());
+        customer.setLastName(request.getLastName());
         customer.setHasValidLicence(request.getIsValidLicence());
         customer.setBirthDate(request.getBirth_date());
 
@@ -31,10 +36,15 @@ public class CustomerManager implements CustomerService {
 
     @Override
     public void delete(DeleteCustomerRequest deleteCustomerRequest) {
-        Customer customerToDelete = customerRepository.findById(deleteCustomerRequest.getCustomerId()).orElseThrow();
+        int customerId = deleteCustomerRequest.getCustomerId();
+
+        if (customerId <= 0) {
+            throw new IllegalArgumentException("Customer ID must be a positive number greater than zero.");
+        }
+
+        Customer customerToDelete = customerRepository.findById(customerId).orElseThrow();
         customerRepository.delete(customerToDelete);
     }
-
     @Override
     public void update(UpdateCustomerRequest updateCustomerRequest) {
         Customer customerToUpdate = customerRepository.findById(updateCustomerRequest.getCustomerId()).orElseThrow();
@@ -45,4 +55,23 @@ public class CustomerManager implements CustomerService {
 
         customerRepository.saveAndFlush(customerToUpdate);
     }
+
+    @Override
+    public List<Customer> getCustomerByFirstName(String firstName) {
+        return customerRepository.findByFirstNameStartingWith(firstName);
+    }
+
+    @Override
+    public List<GetListCustomerResponse> getByCustomerFirstNameDto(String firstName) {
+        return customerRepository.findByFirstNameStartingWith(firstName).stream().map(customer -> {
+            return new GetListCustomerResponse(customer.getFirstName(), customer.getLastName(), customer.getHasValidLicence());
+        }).toList();
+    }
+
+    @Override
+    public List<GetListValidLicenceCostumerResponse> getByValidLicenceDto(Boolean hasValidLicence) {
+        return customerRepository.findByValidLicence(hasValidLicence);
+    }
+
+
 }
